Revert "UFS: add a sysbus UFS device"
This reverts commit f491effd6b198bf67d655347f453caf0ac7768f9.
