+++ hw/ufs/Kconfig
@@ -1,8 +1,4 @@
-config UFS
-    bool
-    default y if SCSI
-
 config UFS_PCI
     bool
     default y if PCI_DEVICES
-    depends on PCI && UFS
+    depends on PCI
