+++ hw/ufs/meson.build
@@ -1,2 +1,1 @@
-system_ss.add(when: 'CONFIG_UFS', if_true: files('ufs.c', 'lu.c'))
-system_ss.add(when: 'CONFIG_UFS_PCI', if_true: files('ufs_pci.c', 'lu_pci.c'))
+system_ss.add(when: 'CONFIG_UFS_PCI', if_true: files('ufs.c', 'lu.c'))
